refactor decorators & helpers locations #35
