+++ src/interactive_pipe/helper/_private.py
@@ -1,0 +1,1 @@
+registered_controls_names = []
